pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "basicstfm"
version = "0.1.0"
description = "A BasicTS-inspired framework for spatio-temporal foundation model research."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "BasicSTFM Contributors" }
]
dependencies = [
  "numpy>=1.23",
  "PyYAML>=6.0",
  "torch>=2.1",
  "tqdm>=4.66"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.4"
]

[project.scripts]
basicstfm = "basicstfm.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"
